refactor: move Warning type and model to separate files
